Treat MAX_CREATES_PER_MINUTE as a float.
The MAX value is an integer, but was being divided by 60, and so
was producing an integer result.  Therefore, if MAX was < 60,
the result is 0 and no new metrics are created after the initial
MAX is hit.  If MAX is >= 60, new metrics will be created, though
not at exactly the rate that would be expected.

The TokenBucket expects floating point values, so force the
division to be floating point.

Fixes issue #55.