Do not force logging to stdlog when not daemonizing
Not forking in the background helps carbon play nice with
process supervisors, in which case we'd rather stick to
logging in the logdir when available.

If there are cases where debug is not wanted but logs are
expected to pop up on stdout, I guess a --stdout-log option
can be added.