Initial code to split maintenance task by C* node
Added --rollup-targets to accept a CSV list of C* nodes to perform
the rollup for. When specified we read the Data Nodes within the token
ranges owned by the C* node and do the rollup on them. This can be used
to run the process in parallel.

TODO: pycassa does not expose describe_token_map() which is used to get the
one to one token to node map. i.e. the tokens assigned to a single node. It
exposes describe_ring() which maps from token ranges to all replicas.
Initally using the describe_ring for proof of concept, will want to use both.
Get the tokens for a node from describe_token_map and then find the ranges
that end with the tokens from describe_ring()

See comments in acquia-maintenance.py

Example execution

  ./acquia-maintenance.py --configdir=/opt/graphite/conf/carbon-daemons/writer/ --dc-name=datacenter1 --serverlist=172.16.1.18 --rollup-targets=172.16.1.18 acquia_rollup