aggregator.receiver: correct test for unaggregated diagnostic
The test added in commit 4d87d706a4d3cf9dad33761cb1bcece5644d95a6 was to
avoid clobbering an incoming metric that was changed by aggregation
(viz. an output aggregation matches the input metric).  The diagnostic
added in 4424e467d04065061057754a460e0a964c4dd2e7 mistakenly suggests
that the metric was not aggregated in this case.  In fact, the original
situation is highly unlikely to occur so in most cases the original
metric will be forwarded as expected, but the diagnostic should be
emitted only (a) when such diagnostics are requested and (b) when the
metric did not contribute to an aggregation.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>