Skip caching of aggregation transformation output
Reported in issue #185 via @dkulikovsky
Each aggregation rule caches transformations to avoid unneeded regex
matches. In the case of many rules, each rule will maintain a cache for
ever metric it sees, whether or not that metric is matched. This
manifests for some users as a memory leak

Instead, compromise by not caching the negative result, storing only
metric transformations that actually match have an output