Clear buffers when they are reset.
The aggregator can use multiple buckets in an aggregation computation. It marks
each bucket as inactive after the computation, but buckets can be reactivated
if new points are added - this can occur if the aggregation computation timing
doesn't align with the bucket boundaries. If that happens, then points will be
re-used between computations, leading to inflated or otherwise incorrect
values. Cleaning buckets after they are used ensures that old data is never
used again even though the bucket may be reused.