Add 'tuned' cache strategy in carbon-cache
Several steps are executed in this strategy in a looping way, that allow to choose which queues to flush
Each step can be controlled in duration or in number of queues to process (absolute or percent) and can be
individually disabled
step #1 : flush queues sorted by metric number (as the 'sorted' strategy). However, this allow to stop
          flushing the list before the end and let's process only queues with lots of metrics.
          Choosing these queues are usefull to decrease I/O, as this maximize the number of datapoints
          to write per whisper udpate_many()
          The 'max' strategy works in the same way, however, it is too CPU intensive with lot of queues
          This strategy tends to keep in memory queues with slow update rate, so it has to be used with
          the persitant cache procedure to prevent data lost if stopping process.
          CACHE_WRITE_TUNED_STRATEGY_LARGEST control this step
            CACHE_WRITE_TUNED_STRATEGY_LARGEST = 100% is equivalent to the 'sorted' strategy
            CACHE_WRITE_TUNED_STRATEGY_LARGEST = 1    is equivalent to the 'max' strategy
step #2 : queues to flush are choosen in a random way (as the 'naive' strategy).
          CACHE_WRITE_TUNED_STRATEGY_RANDOM control this step
step #3 : flush queues sorted by metric age. This step allow to control the time queues are keep in cache
          without being flushed (that could create problems in aggregation computation in whisper).
          CACHE_WRITE_TUNED_STRATEGY_OLDEST control this step
step #4 : flush queues by name. The configuration file 'flushlist.conf' contains regexp that describes
          queues to flush
This strategy could be usefull to use carbon cache with lots of metrics without a high perfomance
storage backend. Using this strategy with a low MAX_UPDATES_PER_SECOND allow to process lots of metrics
and keep the system 'alive' (CPU, I/O...)
This require however much more memory as metrics are accumulated in memory for long time.