Add CACHE_LOW_WATERMARK_PCT setting.
Gives user control of what percentage the CACHE_SIZE_LOW_WATERMARK value
should be of MAX_CACHE_SIZE, which in turn allows to fine tuning when
cacheSpaceAvailable should be triggered if USE_FLOW_CONTROL=True.

In a site where there is a high volume of metrics incoming (500k/minute)
having a higher percentage could result in the carbon-cache service stalling
when the drain throughput falls as it will always hitting cacheTooFull every
couple of seconds.  Reducing the percentage gives it more time to recover.