Route generated points to a specific pipeline
Generated points should not be redirected to the normal
pipeline as this creates infinite loops if, for example, the
name of the aggregated metric is the same as the name of the
original metrics. Instead setup a pipeline that just relays
generate points to the next hop.

This is effectively the behavior of carbon 0.9.*.