The example config and default value were not being set.
Setting these means the proper cast is set when configs are run and this doesn't error out.

Remove example config for some code that does *not* exist anymore.  It was for a config value that would have changed the frequency of the aggregation runs, but this way exists now, even if it is less powerful.