Add DYNAMIC_ROUTER setting
The idea is that you want your consistent hashing ring to
adapt itself when destinations are down. This change will
remove destinations from the router when down and add them
back when up.

An additional buffer is created to store points when no
destination is up, the receiving socket is also paused in this
case to apply some back-pressure.

When a node is removed from the router its points are re-injected
to carbon to go through the new route.