Fix carbon line client protocol so it can send correct data
Currently when `DESTINATION_PROTOCOL` is set to `line` for carbon-relay.
It sends incorrect data to the destination. The order of metric value
and timestamp are changed from the source, for example, it turns
"test.mytest 1 1490397060" into "test.mytest 1490397060 1".

`CarbonLineClientProtocol` was introduced in a previous pull request #601
to support pluggable client protocols along with a number of other
changes. The incorrect order takes place in its method
`_sendDatapointsNow()`.

This fix will correct the order of segments, so metrics path, metric
value and timestamp will be placed in the right order. An unit test is
also added to ensure carbon line client protocol will send correct data
in the future.