Backporting fnv1a_hash ring from graphite-web
It's not a full support of fnv1a hashing in carbon, it's only hashing part. I want to use that in carbonate for rebalance of fnv clusters.
Completely not sure about python 3.x - my tests shows that fnv1a function accepts unicode and non unicode strings with same results but that need to be tested.