Fix carbon cache locking
Add lock when calling choose_item(), this seemed to cause the issues in #822 #815.
Remove previously added locks to counts and watermarks as they deadlocked with
the new lock.
Improve locking in store(), it would lock _after_ checking if the timestamp is in
the cache. The cache could have been altered between checking and locking.