Bucketmax write strategy
Bucketmax writes the metric with the most datapoints just as max does but uses
a different approach to determine which metric has the most datapoints.

At each write operation 'max' loops once over the list of metrics and picks the metric
with most datapoints.

'bucketmax' stores all metrics in a list where each index represents the number of
datapoints in the specific metrics. For example `buckets[5]` contains a list of all
metrics which have 6 datapoints. At each write operation 'bucketmax' selects the
first metric from the largest bucket.

The time complexity of 'max' is linear, 'bucketmax' is near constant time but runs
on every ingestion where 'max' only runs on every write. There is also some overhead
in resizing the buckets list. I've done some limited testing and 'bucketmax'' CPU
usage was almost half of that of 'max'. The constraint of bucketmax is that the number
of datapoints per metrics must be relatively low compared to the total number of
unique metrics (which I think is the most common use-case in Graphite usage). I don't
know where the tipping point is but if one has for example 10 metrics with each 100000
datapoints I would assume 'bucketmax' would perform worse than 'max'.