Update rewrite-rules.conf.example - explicitly state aggregator
Explicitly state and relate rewrite-rules.conf to carbon-aggregator so that this info is available to the user in-situ in the example. There is ambiguity in the phrase "as soon as they are received" this can confusing because carbon-relay could be seen as where they are being received.

Related to https://github.com/graphite-project/graphite-web/pull/2845 which tries to make it clearer in the docs as well.